dmidecode: Drop the CPUID exception list
Back in 2003, I had a system where the CPU type was not set. I added
a quirk so that it would still be recognized as x86, and the CPUID
could be decoded.

A few more exceptions where added over the years, but in effect, the
list was last modified in 2008.

Having such an exception list isn't actually a good idea, for the
following reasons:
 * It requires endless maintenance work if we want to keep it
   up-to-date.
 * It adds some (admittedly minimal) burden to the sane systems.
 * If we were to add more entries to the exception list, it wouldn't
   scale well (linear algorithmic complexity). This could be improved
   but at the cost of more complex code.
 * It sends the wrong message to the hardware manufacturers ("You can
   get things wrong, we'll add a workaround on our side.")

Therefore I would like to get rid of this exception list. Doing so
has the nice side effect of simplifying the code and making the
binary smaller.

If anyone really needs the CPUID information on such non-compliant
systems, there are other ways to retrieve it, such as lscpu or
/proc/cpuinfo.

Signed-off-by: Jean Delvare <jdelvare@suse.de>