dmidecode: Write the whole dump file at once
When option --dump-bin is used, write the whole dump file at once,
instead of opening and closing the file separately for the table
and then for the entry point.

As the file writing function is no longer generic, it gets moved
from util.c to dmidecode.c.

One minor functional change resulting from the new implementation is
that the entry point is written first now, so the messages printed
are swapped.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Reviewed-by: Jerry Hoemann <jerry.hoemann@hpe.com>