dmidecode: Fix flag mask in dmi_slot_characteristics
When slot characteristics flags were added, the mask to test if at
least one flag is set wasn't updated. This could result in slot
characteristics being listed as "None" for slots with only new flags
set.

Fix this by dropping the mask for the second slot characteristics
byte completely. It seems safe to assume that all the remaining bits
of this byte will be used to future characteristic flags.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Fixes: f7d236dd5c4d ("dmidecode: Remaining support for SMBIOS 3.2.0")
Fixes: 4304d3aca801 ("Add new slot characteristics2 from SMBIOS spec 3.4.0")