dmidecode: Update the sockets in Processor Update section
When running the dmidecode command to check SMBIOS Processor Information
(type 4), the "Upgrade" for CPU in the result shows "<OUT OF SPEC>",
because the latest version of the dmidecode source code doesn't cover
the sockets beyond 50h in the Processor Information — Processor Upgrade
section of SMBIOS version 3.8. So, add the parts of the SMBIOS 3.8
Processor Information — Processor Upgrade section beyond socket 50h to
the end of "dmi_processor_upgrade" in the source code, and update the
conditional statements to include the current last value, 57h.

Signed-off-by: Jessie Chen <ychen39@lenovo.com>
Signed-off-by: Jean Delvare <jdelvare@suse.de>