dmidecode: Update info in Processor Family field
When running the dmidecode command to check SMBIOS Processor Information
(type 4), the "Family" for CPU in the result shows "<OUT OF SPEC>",
because the latest version of the dmidecode source code doesn't cover
the values for the Processor Information beyond 300h in the
Processor Information — Processor Family section of SMBIOS version 3.8.
So, add the parts of the SMBIOS 3.8 Processor Information — Processor Family
section beyond value 300h to the end of "0x271" in the source code.

Signed-off-by: Jessie Chen <ychen39@lenovo.com>
Signed-off-by: Jean Delvare <jdelvare@suse.de>